Also fix for calendar view
